9lab

+++ app/router.py
@@ -16,14 +16,25 @@ def get_repo():
 
 @router.post("/", response_model=UserRead, status_code=201)
 def create_user(data: UserCreate, repo: UserRepository = Depends(get_repo)):
+    lock_key = f"lock:user:create:{data.username}"
+
+    # INFO: Redis Lock
+    acquired = cache_service.client.set(lock_key, "locked", ex=30, nx=True)
+
+    if not acquired:
+        raise HTTPException(
+            status_code=423, detail="Пользователь с таким именем уже создается"
+        )
+
     try:
         user = repo.create(data)
         cache_service.delete_by_pattern("wp:users:list:*")
         return UserRead.model_validate(user)
-
     except Exception as e:
         print(f"Error: {e}")
         raise HTTPException(409, "Конфликт: пользователь уже существует")
+    finally:
+        cache_service.client.delete(lock_key)
 
 
 @router.get("/", response_model=PaginatedUsers)
@@ -78,8 +89,6 @@ def delete_user(
     cache_service.delete_by_pattern("wp:users:list:*")
     cache_service.delete(f"wp:users:profile:{user_id}")
 
-    # --- ПРОФИЛЬ ПОЛЬЗОВАТЕЛЯ ---
-
 
 profile_router = APIRouter(prefix="/profile", tags=["Profile"])
 
@@ -101,29 +110,24 @@ def update_profile(
     current: dict = Depends(get_current_user),
     repo: UserRepository = Depends(get_repo),
 ):
-    """Обновление профиля (установка аватара)"""
     user_id = current["user_id"]
     user = repo.get_by_id(user_id)
 
     if not user:
         raise HTTPException(404, "Пользователь не найден")
 
-    # 1. Если передали ID файла для аватара, проверяем его!
     if data.avatarFileId:
         from app.models import File
 
-        # Ищем файл в БД
         db_file = File.objects(id=data.avatarFileId, deleted_at=None).first()
         if not db_file:
             raise HTTPException(404, "Файл не найден")
 
-        # САМАЯ ВАЖНАЯ ПРОВЕРКА - принадлежит ли файл этому юзеру?
         if str(db_file.user.id) != user_id:
             raise HTTPException(403, "У вас нет прав на использование этого файла")
 
         user.avatar_file_id = db_file.id
 
-    # 2. Обновляем остальные текстовые поля (если они есть)
     if data.username:
         user.username = data.username
     if data.os:
@@ -133,7 +137,6 @@ def update_profile(
 
     user.save()
 
-    # 3. Инвалидируем (сбрасываем) кеш профиля и списка юзеров
     cache_service.delete(f"wp:users:profile:{user_id}")
     cache_service.delete_by_pattern("wp:users:list:*")
